+++ src/Correlate.Core/RootActivity.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Diagnostics;
+using Correlate.Extensions;
 using Microsoft.Extensions.Logging;
 
 namespace Correlate
